SDK-632 -- Disable Tracking option does not persist throughout multiple sessions.

+++ BranchSDK/src/BranchIO/Util/UnixStorage.cpp
@@ -12,7 +12,7 @@ using namespace Poco::Util;
 
 namespace BranchIO {
 
-UnixStorage&
+IStorage&
 UnixStorage::instance() {
     static UnixStorage _instance;
     return _instance;
@@ -37,19 +37,36 @@ UnixStorage::has(const std::string& key, Scope scope) const {
     return getConfig(scope)->has(key);
 }
 
-bool
-UnixStorage::get(const std::string& key, std::string& value, Scope scope) const {
+std::string
+UnixStorage::getString(const std::string& key, const std::string& defaultValue, Scope scope) const {
     Mutex::ScopedLock _l(_mutex);
     StoragePtr config(getConfig(scope));
-    if (!config->has(key)) return false;
-    value = config->getString(key);
-    return true;
+    if (!config->has(key)) return defaultValue;
+
+    return config->getString(key);
 }
 
 IStorage&
-UnixStorage::set(const std::string& key, const std::string& value, Scope scope) {
+UnixStorage::setString(const std::string& key, const std::string& value, Scope scope) {
     Mutex::ScopedLock _l(_mutex);
     getConfig(scope)->setString(key, value);
+    flush(scope);
+    return *this;
+}
+
+bool
+UnixStorage::getBoolean(const std::string& key, bool defaultValue, Scope scope) const {
+    Mutex::ScopedLock _l(_mutex);
+    StoragePtr config(getConfig(scope));
+    if (!config->has(key)) return defaultValue;
+
+    return config->getBool(key, false);
+}
+
+IStorage&
+UnixStorage::setBoolean(const std::string& key, bool value, Scope scope) {
+    Mutex::ScopedLock _l(_mutex);
+    getConfig(scope)->setBool(key, value);
     flush(scope);
     return *this;
 }
@@ -80,7 +97,7 @@ UnixStorage::clear(Scope scope) {
     }
 
     File file(getPath(scope));
-    file.remove(true);  // recursively removes the directory
+    if (file.exists()) file.remove(true);  // recursively removes the directory
 
     return *this;
 }
